Create main3.cs
This repository contains a C# console application that allows users to remove a specified digit from a given natural number.

Features
Prompts the user to input a natural number.
Prompts the user to input a digit to be removed from the number.
Validates user input to ensure correctness.
Removes all instances of the specified digit from the number.
Displays the resulting number after the digit has been removed.